discord-bot/requirements.txt
discord.py
python-dotenv
fastapi
uvicorn
pydantic
